Use per-sentence lemmas rather than per-article lemmas.
The reasoning is, we care about the frequency of the lemma in general but with per-article we lose the information of how often it appears in the article itself. (i.e. a lemma that appears 20 times in one article is evaluated the same as a lemma that appears one time in one article).